VER-0.01-.50: MeatSeekerMissileLauncher updated to use new prefab and sound triggers
