--- Assets/Audio/TechSfx/ConstructSfx/hit07.wav.meta
@@ -1,22 +1,0 @@
-fileFormatVersion: 2
-guid: 38eb3ce977e73a54a8b2c638bc63509e
-timeCreated: 1540231160
-licenseType: Store
-AudioImporter:
-  serializedVersion: 6
-  defaultSettings:
-    loadType: 0
-    sampleRateSetting: 0
-    sampleRateOverride: 44100
-    compressionFormat: 1
-    quality: 1
-    conversionMode: 0
-  platformSettingOverrides: {}
-  forceToMono: 0
-  normalize: 1
-  preloadAudioData: 1
-  loadInBackground: 0
-  3D: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
