--- Assets/Prefabs/Weapons/MeatSeekingMissles.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 48f67ac72670c824db7182ffd088c2d7
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
