+++ Assets/Scenes/MainCorridorScene.scenetemplate
@@ -545,7 +545,7 @@ MonoBehaviour:
     instantiationMode: 1
   - dependency: {fileID: 2100000, guid: dfcc4132567ccff43ae181c20fddbd34, type: 2}
     instantiationMode: 0
-  - dependency: {fileID: 7362472295509532169, guid: 48f67ac72670c824db7182ffd088c2d7, type: 3}
+  - dependency: {fileID: 2800000, guid: 854b25bc08516364da5a786b1cad247e, type: 3}
     instantiationMode: 0
   - dependency: {fileID: 2100000, guid: 28f3f954bc572de43a3bf96b81c9d1cf, type: 2}
     instantiationMode: 0
@@ -567,8 +567,6 @@ MonoBehaviour:
     instantiationMode: 1
   - dependency: {fileID: 8300000, guid: 92d3259b6a9732e4ba3fb95331b2c95d, type: 3}
     instantiationMode: 1
-  - dependency: {fileID: 2800000, guid: 854b25bc08516364da5a786b1cad247e, type: 3}
-    instantiationMode: 0
   - dependency: {fileID: 7078411352450516573, guid: 9d9caed540f78cc458274ab63a03ec52, type: 3}
     instantiationMode: 0
   - dependency: {fileID: 8300000, guid: 448511260f8777a49bc36049171e0797, type: 3}
@@ -1373,6 +1371,8 @@ MonoBehaviour:
     instantiationMode: 0
   - dependency: {fileID: 2100000, guid: 723956fd057603a4e92c683ce37a679f, type: 2}
     instantiationMode: 0
+  - dependency: {fileID: 3981800169902183240, guid: 61a722aa53457af4f85ac6aefd3f64a3, type: 3}
+    instantiationMode: 0
   - dependency: {fileID: 1212038522573524, guid: 6736c17ffc1bf784eb6849ac6ed3ead1, type: 3}
     instantiationMode: 0
   - dependency: {fileID: 11400000, guid: b737a817c88ddf547884acd11313d710, type: 2}
@@ -1507,6 +1507,8 @@ MonoBehaviour:
     instantiationMode: 1
   - dependency: {fileID: 2800000, guid: 06907d08092c41f4aa8b5b3cbe5f9fa2, type: 3}
     instantiationMode: 0
+  - dependency: {fileID: 8300000, guid: 38eb3ce977e73a54a8b2c638bc63509e, type: 3}
+    instantiationMode: 1
   - dependency: {fileID: 8300000, guid: 62879db7f7abc1c47b0ed08791e7eaf4, type: 3}
     instantiationMode: 1
   - dependency: {fileID: 11400000, guid: 7a21f24703be1924cbd2a551ec5c8111, type: 2}
