+++ Assets/Scripts/Weapons/Projectiles/MeatSeeker.cs
@@ -1,8 +1,8 @@
 using Controllers;
-using Controllers.Pools;
 using Damagers;
 using UnityEngine;
 using UnityEngine.Splines;
+using Weapons.SpecificWeapons;
 
 namespace Weapons.Projectiles
 {
@@ -21,27 +21,24 @@ namespace Weapons.Projectiles
 
         private void Detonate()
         {
-            var exp = ProjectilePoolManager2.poolProj.projPools[pid].GetPooledGameObject();
+            var exp = MeatSeekingMissileLauncher.Instance.explosionPool.GetPooledGameObject();
             var damager = exp.GetComponent<EExplosionDamager>();
             damager.damage = damage;
             damager.blastRadiusCollider.radius = blastRadius;
             exp.gameObject.transform.position = transform.position;
             exp.SetActive(true);
             GenericShakeController.Instance.ShakeWeakStrongViolent("strong", transform);
+            MeatSeekingMissileLauncher.Instance.juiceManager.TriggerFeedback(GenericJuiceManager.FeedbackType.Death);
             animator.Restart(false);
-            missileTrack.SetActive(false); }
+            missileTrack.SetActive(false); 
+        }
 
         private void OnDisable()
         {
-            var exp = ProjectilePoolManager2.poolProj.projPools[pid].GetPooledGameObject();
-            var damager = exp.GetComponent<EExplosionDamager>();
-            damager.damage = damage;
-            damager.blastRadiusCollider.radius = blastRadius;
-            exp.gameObject.transform.position = transform.position;
-            exp.SetActive(true);
-            GenericShakeController.Instance.ShakeWeakStrongViolent("strong", transform);
-            animator.Restart(false);
-            missileTrack.SetActive(false);
+            MeatSeekingMissileLauncher.Instance.juiceManager.StopFeedback(GenericJuiceManager.FeedbackType.Death);
+            MeatSeekingMissileLauncher.Instance.juiceManager.StopFeedback(GenericJuiceManager.FeedbackType.Firing);
+            // idle is the radar ping
+            MeatSeekingMissileLauncher.Instance.juiceManager.StopFeedback(GenericJuiceManager.FeedbackType.Idle);
         }
     }
 }
