--- Assets/Scripts/Weapons/SpecificWeapons/MissileWeapon.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 82691a055a93cc449b4cb5b8f6aa3022
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
